Move ExponentialBackoff to vespajlib
